pyproject.toml
[project]
name = "imdb-qa-pgvector-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.23",
    "langchain-openai>=0.3.14",
    "langchain-postgres>=0.0.14",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
